Move metrics_grabbert to test/e2e
