+++ pkg/BUILD
@@ -101,7 +101,6 @@ filegroup(
         "//pkg/kubemark:all-srcs",
         "//pkg/labels:all-srcs",
         "//pkg/master:all-srcs",
-        "//pkg/metrics:all-srcs",
         "//pkg/printers:all-srcs",
         "//pkg/probe:all-srcs",
         "//pkg/proxy:all-srcs",
