+++ test/e2e/BUILD
@@ -19,12 +19,12 @@ go_test(
     tags = ["automanaged"],
     deps = [
         "//pkg/client/clientset_generated/clientset:go_default_library",
-        "//pkg/metrics:go_default_library",
         "//test/e2e/api:go_default_library",
         "//test/e2e/autoscaling:go_default_library",
         "//test/e2e/cluster-logging:go_default_library",
         "//test/e2e/extension:go_default_library",
         "//test/e2e/framework:go_default_library",
+        "//test/e2e/metrics:go_default_library",
         "//test/e2e/perf:go_default_library",
         "//test/e2e/scheduling:go_default_library",
         "//test/e2e/storage:go_default_library",
@@ -128,7 +128,6 @@ go_library(
         "//pkg/kubelet/apis/stats/v1alpha1:go_default_library",
         "//pkg/kubelet/pod:go_default_library",
         "//pkg/master/ports:go_default_library",
-        "//pkg/metrics:go_default_library",
         "//pkg/quota/evaluator/core:go_default_library",
         "//pkg/util/exec:go_default_library",
         "//pkg/util/logs:go_default_library",
@@ -140,6 +139,7 @@ go_library(
         "//test/e2e/framework/ginkgowrapper:go_default_library",
         "//test/e2e/generated:go_default_library",
         "//test/e2e/manifest:go_default_library",
+        "//test/e2e/metrics:go_default_library",
         "//test/e2e/perf:go_default_library",
         "//test/e2e/scheduling:go_default_library",
         "//test/e2e/upgrades:go_default_library",
@@ -228,6 +228,7 @@ filegroup(
         "//test/e2e/framework:all-srcs",
         "//test/e2e/generated:all-srcs",
         "//test/e2e/manifest:all-srcs",
+        "//test/e2e/metrics:all-srcs",
         "//test/e2e/perf:all-srcs",
         "//test/e2e/perftype:all-srcs",
         "//test/e2e/scheduling:all-srcs",
