+++ test/e2e/framework/BUILD
@@ -72,7 +72,6 @@ go_library(
         "//pkg/kubelet/sysctl:go_default_library",
         "//pkg/kubelet/util/format:go_default_library",
         "//pkg/master/ports:go_default_library",
-        "//pkg/metrics:go_default_library",
         "//pkg/ssh:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/exec:go_default_library",
@@ -84,6 +83,7 @@ go_library(
         "//plugin/pkg/scheduler/schedulercache:go_default_library",
         "//test/e2e/framework/ginkgowrapper:go_default_library",
         "//test/e2e/manifest:go_default_library",
+        "//test/e2e/metrics:go_default_library",
         "//test/e2e/perftype:go_default_library",
         "//test/utils:go_default_library",
         "//vendor/github.com/aws/aws-sdk-go/aws:go_default_library",
