+++ test/e2e/metrics/BUILD
