+++ test/e2e/metrics/controller_manager_metrics.go
