+++ test/e2e_node/BUILD
@@ -35,10 +35,10 @@ go_library(
         "//pkg/kubelet/apis/stats/v1alpha1:go_default_library",
         "//pkg/kubelet/metrics:go_default_library",
         "//pkg/kubelet/remote:go_default_library",
-        "//pkg/metrics:go_default_library",
         "//pkg/util/procfs:go_default_library",
         "//test/e2e/common:go_default_library",
         "//test/e2e/framework:go_default_library",
+        "//test/e2e/metrics:go_default_library",
         "//test/e2e/perftype:go_default_library",
         "//test/e2e_node/perftype:go_default_library",
         "//vendor/github.com/golang/glog:go_default_library",
@@ -109,10 +109,10 @@ go_test(
         "//pkg/kubelet/images:go_default_library",
         "//pkg/kubelet/metrics:go_default_library",
         "//pkg/kubelet/types:go_default_library",
-        "//pkg/metrics:go_default_library",
         "//pkg/security/apparmor:go_default_library",
         "//test/e2e/common:go_default_library",
         "//test/e2e/framework:go_default_library",
+        "//test/e2e/metrics:go_default_library",
         "//test/e2e_node/services:go_default_library",
         "//test/e2e_node/system:go_default_library",
         "//test/utils:go_default_library",
